+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/pcm2markov/MarkovTransformation.java
@@ -11,7 +11,7 @@ import org.palladiosimulator.reliability.helper.MarkovFailureTypeHelper;
 import org.palladiosimulator.reliability.markov.MarkovChain;
 import org.palladiosimulator.reliability.solver.MarkovSolver;
 import org.palladiosimulator.reliability.solver.sensitivity.MarkovSensitivity;
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
 import org.palladiosimulator.solver.visitors.UsageModelVisitor;
 import org.palladiosimulator.pcm.usagemodel.UsageScenario;
