+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/runconfig/RunPCMReliabilityAnalysisJob.java
@@ -5,7 +5,7 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.palladiosimulator.analyzer.workflow.core.blackboard.PCMResourceSetPartition;
 import org.palladiosimulator.analyzer.workflow.jobs.LoadPCMModelsIntoBlackboardJob;
 import org.palladiosimulator.reliability.solver.pcm2markov.Pcm2MarkovStrategy;
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
 import org.palladiosimulator.solver.transformations.SolverStrategy;
