+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/sensitivity/MultiSensitivity.java
@@ -3,7 +3,7 @@ package org.palladiosimulator.reliability.solver.sensitivity;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 
 /**
  * Provides a possibility to define multiple sensitivity parameters to be defined within one
